take into account caching | hash for boom-template takes into account rocket

+++ .circleci/prepare-for-rtl-build.sh
@@ -5,33 +5,35 @@
 # turn echo on and error on earliest command
 set -ex
 
-# clone the boom-template repo
-cd $HOME
+if [ ! -d "$HOME/boom-template" ]; then
+    # clone the boom-template repo
+    cd $HOME
 
-# clone boom-template and create the riscv-tools
-git clone --progress --verbose https://github.com/riscv-boom/boom-template.git
-cd boom-template
+    # clone boom-template and create the riscv-tools
+    git clone --progress --verbose https://github.com/riscv-boom/boom-template.git
+    cd boom-template
 
-# init all submodules (according to what boom-template wants)
-./scripts/init-submodules-no-riscv-tools.sh
+    # init all submodules (according to what boom-template wants)
+    ./scripts/init-submodules-no-riscv-tools.sh
 
-# move the pull request riscv-boom repo into boom-template
-rm -rf $HOME/boom-template/boom
-cp -r $HOME/project $HOME/boom-template/boom/
+    # move the pull request riscv-boom repo into boom-template
+    rm -rf $HOME/boom-template/boom
+    cp -r $HOME/project $HOME/boom-template/boom/
 
-# get boom specific rocket-chip version
-echo "Checking out rocket-chip with hash: $(cat boom/ROCKETCHIP_VERSION)"
-cd rocket-chip
-git fetch
-git checkout $(cat ../boom/ROCKETCHIP_VERSION)
+    # get boom specific rocket-chip version
+    echo "Checking out rocket-chip with hash: $(cat boom/ROCKETCHIP_VERSION)"
+    cd rocket-chip
+    git fetch
+    git checkout $(cat ../boom/ROCKETCHIP_VERSION)
 
-echo "Initialize final submodules"
-git submodule update --init --recursive
+    echo "Initialize final submodules"
+    git submodule update --init --recursive
 
-# extra patches
-# TODO: Remove FIRRTL patch in next rocket-bump (fixes const prop issue)
-git -C firrtl checkout 9535e03020c6e654dae3ce7e95f4d8649405ce3d
+    # extra patches
+    # TODO: Remove FIRRTL patch in next rocket-bump (fixes const prop issue)
+    git -C firrtl checkout 9535e03020c6e654dae3ce7e95f4d8649405ce3d
 
-# make boom-template verilator version
-cd ../verisim
-make verilator_install
+    # make boom-template verilator version
+    cd ../verisim
+    make verilator_install
+fi
